Refactor open_zip and create_zip. Many new test cases.
open_zip now supports read and write.

create_zip now only does 1-line zip creation from Pairs, Dict etc...

open_zip now returns ZipArchive <: Associative.
This wraps Reader and Writer.
ZipArchive has as cache to support reading and writing an archive.
e.g. if the user opens an existing archive, then adds a file, the
existing files are read in to the cache, then the reader is closed
and a writer is created to write everyting out to the output archive.
If the user only reads or only writes then the cache is not used and
the previous memory-efficient behaviour is retained where one file
is handled at a time. If reads and writes are mixed, the cache is
activated and the whole content is stored in ram. This could be made
more efficient later by stroring compressed data in the cache to avoid
decompressing/recompressing.

ZipFile.jl

type Reader and Writer:
Add _close_io::Bool flag to tell close() if it is allowed to close
the underlying stream (sometimes the caller wants it left open).

Add rewind(f::ReadableFile) to allow a ReadableFile to be read more
than once (not exported).

Fix bug in write(f::WritableFile, p::Ptr, nb::Integer) where a zero
length write would cause an error inside _zio.