Enable the UTF-8 bit flag by default
Following the ZIP Format Specification, when using UTF-8 encoded strings
for the filenames, it is recommended to set the Bit 11 of the general
purpose bit flag (see sections 4.4.4 and APPENDIX D of the ZIP
specification v6.3.9). This change is required to produce zip files
containing unicode filenames that are compatible with other zip tools
and libraries (such as Python's zipfile module).