Add OS creation information and default file attributes
Set the "version made by" field depending on the creating OS (currently
either Windows on generic Unix) and add default file attributes for the
files when being created on *nix. These changes are needed for proper
UTF-8 support since some zip tools assume that if the creation OS is
MSDOS (the value being set before this commit) then the non-ASCII
characters are encoded using Code Page 437 and not UTF-8 (even though we
already set the UTF-8 encoding flag these tools seem to ignore it for
MSDOS). This is for example the case for the "unzip" tool by INFO-Zip
found on most Linux distros.