Bugfix getting garbage collected
This commit is an attempt to address
github.com/fhs/ZipFile.jl/issues/14. It adds to ReadableFiles references
to parent Reader to ensure that a Reader can't get garbage collected for
as long as the children ReadableFiles are around.

Possibly reader::Vector{Reader} would be cleaner, but it appears that
mutually-circular type declarations aren't possible yet(?), see
github.com/JuliaLang/issues/269.